requirements.txt
tensorflow-cpu
matplotlib
numpy
pillow
opencv-python
jupyterlab
flask
flask-cors
waitress